pyproject.toml
[tool.poetry]
name = "ai-comrades-template"
version = "0.1.0"
description = "Template project for other AI.Comrades"
authors = ["alexbuyan <alexbuyan.dev@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.13"
crewai = "^0.11.0"
python-dotenv = "1.0.0"
crewai-tools = "^0.0.16"
langchain = "^0.1.12"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"